Tolerate project declarations in text blocks
It is possible for buildscripts to contain project('...') declarations
in locations other than the dependencies {} and buildscript {} blocks.
For example, a plugin may support arbirary syntax in a block like this:

myVendorBlock {
    applicableProjects = [ project(':foo:bar') ]
}

The Gradle Dependencies Sorter won't know how to sort such thing, since
it won't have knowledge of vendor syntax. However, the sorter should
just skip these things instead of failing with a parse error.

Resolves: https://github.com/square/gradle-dependencies-sorter/issues/36