Change resource name to avoid clobbering with other plugins.
This could happen because the original name was very simply 'version.txt', which was also present from the Sonar plugin.
Due to the nature of classloading, both this plugin's resource and the Sonar plugin's resource were available, and the
latter was chosen instead. The simple solution was to namespace version.txt: it is now com-squareup-sort-version.txt.

Resolves #17